+++ docs/releasenotes.md
@@ -42,7 +42,7 @@
 * `BIGDOOR7`, `FIRBLU1` and `SKY1` textures are now displayed correctly.
 * The `vid_motionblur` CVAR now accepts a value between `0%` and `100%`, rather than just `on` or `off`. Its default is `0%`.
 * An `r_skycolor` CVAR has been implemented that allows the player to override the current map’s sky texture and use a solid color for the sky instead. It accepts a value between `0` and `255`, or `none` (the default).
-* The otherwise read-only `wad` CVAR is now reset when the `resetall` CCMD is used.
+* The `wad` CVAR is now reset whenever `reset iwadfolder` or `resetall` is used in the console.
 
 ---